+++ src/core/index.ts
@@ -1,4 +1,5 @@
 export * from './constants.js';
 export * from './interfaces.js';
+export * from './utils/websocket.js';
 export * from './utils/serializers.js';
 export * from './encryption.js';